pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "renova"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [
  { name = "charleain", email = "183850672+charleain@users.noreply.github.com" },
]

dependencies = [
  "fastapi",
  "sqlmodel",
  "httpx",
  "pydantic[email]",
  "jinja2",
  "pyjwt[crypto]",
  "argon2-cffi",
  "python-multipart",
]

[tool.hatch.version]
path = "src/renova/__about__.py"
